// typescript/models/AccountMetadata.ts
/**
 * Railway
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AccountMetadata {
    /**
    * The id of the object.
    */
    'id'?: number;
    /**
    * The id of the object.
    */
    'accountId'?: number;
    /**
    * The id of the object.
    */
    'module'?: string;
    /**
    * The id of the object.
    */
    'key'?: string;
    /**
    * The id of the object.
    */
    'value'?: string;
    /**
    * The creation date of the object.
    */
    'createdAt'?: Date;
    /**
    * The updating date of the object.
    */
    'updatedAt'?: Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "accountId",
            "baseName": "account_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "module",
            "baseName": "module",
            "type": "string",
            "format": ""
        },
        {
            "name": "key",
            "baseName": "key",
            "type": "string",
            "format": ""
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return AccountMetadata.attributeTypeMap;
    }

    public constructor() {
    }
}